requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alembic==1.17.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
appdirs==1.4.4
attrs==25.4.0
banal==1.0.6
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
cfgv==3.4.0
charset-normalizer==3.4.4
click==8.3.0
dataclasses-json==0.6.7
dataset==1.6.2
datasets==4.4.1
dill==0.4.0
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
docstring_parser==0.17.0
evaluate==0.4.6
faiss-cpu==1.13.0
filelock==3.20.0
filetype==1.2.0
frozenlist==1.8.0
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.9.0
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.42.1
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.71.0
groq==0.37.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
identify==2.6.15
idna==3.11
instructor==1.13.0
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langchain==1.1.0
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.1.0
langchain-google-genai==3.2.0
langchain-groq==1.1.0
langchain-huggingface==1.1.0
langchain-openai==1.0.3
langchain-text-splitters==1.0.0
langgraph==1.0.2
langgraph-checkpoint==3.0.0
langgraph-prebuilt==1.0.2
langgraph-sdk==0.2.9
langsmith==0.4.39
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
mypy_extensions==1.1.0
narwhals==2.11.0
nest-asyncio==1.6.0
networkx==3.5
nodeenv==1.9.1
numpy==2.3.4
openai==2.8.1
orjson==3.11.4
ormsgpack==1.11.0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
platformdirs==4.5.0
pre_commit==4.4.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.5
pypdf==6.1.3
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
ragas==0.3.9
referencing==0.37.0
regex==2025.10.23
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.28.0
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.2
scikit-network==0.33.5
scipy==1.16.3
sentence-transformers==5.1.2
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==1.4.54
streamlit==1.51.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
toml==0.10.2
torch==2.9.0
tornado==6.5.2
tqdm==4.67.1
transformers==4.57.1
ty==0.0.1a27
typer==0.20.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
virtualenv==20.35.4
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
